adding weather class Gpio class upgrade add functionnalities for weather alarm
