+++ F3FChrono/chrono/GPIOPort.py
@@ -1,12 +1,11 @@
 # vim: set et sw=4 sts=4 fileencoding=utf-8:
 import threading
 from time import sleep
-
-
-
+from PyQt5.QtCore import pyqtSignal, QObject, QTimer
 import RPi.GPIO as GPIO
 from PyQt5.QtCore import QObject, pyqtSignal
 from F3FChrono.chrono import ConfigReader
+
 
 def statusLED(port, on=True):
     """
@@ -26,7 +25,7 @@ def addCallback(port, fctn, falling=True):
     """
     # TODO: add exception handling
     GPIO.setmode(GPIO.BCM)
-    #GPIO.setwarnings(False)
+    # GPIO.setwarnings(False)
     if falling:
         GPIO.setup(port, GPIO.IN, pull_up_down=GPIO.PUD_UP)
         GPIO.add_event_detect(port, GPIO.FALLING, callback=fctn, bouncetime=500)
@@ -35,76 +34,17 @@ def addCallback(port, fctn, falling=True):
         GPIO.add_event_detect(port, GPIO.RISING, callback=fctn, bouncetime=500)
 
 
-
-class gpioPort(threading.Thread):
+class gpioPort(QTimer):
     def __init__(self, port, duration=200., is_active_low=False, start_blinks=0):
-        super(gpioPort, self).__init__()
-        self.terminated = False
-        self.duration   = duration
-        self.event      = threading.Event()
-        self.port       = port
-        self.activate   = GPIO.HIGH
-        self.deactivate = GPIO.LOW
-
-        GPIO.setmode(GPIO.BCM)
-        GPIO.setup(self.port,GPIO.OUT)
-
-        if is_active_low:
-            self.activate   = GPIO.LOW
-            self.deactivate = GPIO.HIGH
-
-        if start_blinks > 0:
-            self.blink(start_blinks)
-
-        GPIO.output(self.port,self.deactivate)
-
-        self.daemon = True
-        self.event.clear()
-        self.start()
-
-    def blink(self, numbers):
-        if (self.port == ConfigReader.config.conf['buzzer'] and ConfigReader.config.conf[
-            'buzzer_valid'] or
-                self.port == ConfigReader.config.conf['buzzer_next'] and ConfigReader.config.conf[
-                    'buzzer_next_valid']):
-            for i in range(0,numbers):
-                GPIO.output(self.port,self.activate)
-                sleep(self.duration/1000.0)
-                GPIO.output(self.port,self.deactivate)
-                sleep(self.duration/1000.0)
-
-    def check(self, value):
-        self.event.set()
-
-    def run(self):
-        while not self.terminated:
-            # wait until somebody throws an event
-            if self.event.wait(1):
-                if (self.port==ConfigReader.config.conf['buzzer'] and ConfigReader.config.conf['buzzer_valid'] or
-                    self.port == ConfigReader.config.conf['buzzer_next'] and ConfigReader.config.conf['buzzer_next_valid']):
-                    # create rectangle signal on GPIO port
-                    GPIO.output(self.port,self.activate)
-                    sleep(self.duration/1000.0)
-                    GPIO.output(self.port,self.deactivate)
-                    self.event.clear()
-
-        #GPIO.cleanup(self.port)
-        GPIO.cleanup()
-        
-        
-def event_detected(port):
-    print("callback "+str(port))
-
-
-'''class gpioPort(threading.Thread):
-    def __init__(self, port, duration=200., is_active_low=False, start_blinks=0):
-        super(gpioPort, self).__init__()
+        super().__init__()
         self.terminated = False
         self.duration = duration
-        self.event = threading.Event()
         self.port = port
         self.activate = GPIO.HIGH
         self.deactivate = GPIO.LOW
+        self.state = False
+        self.nbevent = 0
+        self.timeout.connect(self.run)
 
         GPIO.setmode(GPIO.BCM)
         GPIO.setup(self.port, GPIO.OUT)
@@ -112,54 +52,58 @@ def event_detected(port):
         if is_active_low:
             self.activate = GPIO.LOW
             self.deactivate = GPIO.HIGH
-
         if start_blinks > 0:
-            self.blink(start_blinks)
+            self.blink("blink", start_blinks)
 
-        GPIO.output(self.port, self.deactivate)
+        self.__deactivate()
 
-        self.daemon = True
-        self.event.clear()
-        self.start()
-
-    def blink(self, numbers):
-        if (self.port == ConfigReader.config.conf['buzzer'] and ConfigReader.config.conf[
-            'buzzer_valid'] or
-                self.port == ConfigReader.config.conf['buzzer_next'] and ConfigReader.config.conf[
-                    'buzzer_next_valid']):
-            for i in range(0, numbers):
-                GPIO.output(self.port, self.activate)
-                sleep(self.duration / 1000.0)
-                GPIO.output(self.port, self.deactivate)
-                sleep(self.duration / 1000.0)
-
-    def check(self, value):
-        self.event.set()
+    def slot_blink(self, mode, numbers, duration=-1):
+        if mode.lower() == "permanent":
+            self.__activate()
+        elif mode.lower() == "stop":
+            self.__deactivate()
+        elif mode.lower() == "blink":
+            self.nbevent = numbers
+            if self.nbevent > 0:
+                if duration == -1:
+                    self.start(self.duration)
+                elif duration > 0:
+                    self.start(duration)
 
     def run(self):
-        while not self.terminated:
-            # wait until somebody throws an event
-            if self.event.wait(1):
-                if (self.port == ConfigReader.config.conf['buzzer'] and ConfigReader.config.conf['buzzer_valid'] or
-                        self.port == ConfigReader.config.conf['buzzer_next'] and ConfigReader.config.conf[
-                            'buzzer_next_valid']):
-                    # create rectangle signal on GPIO port
-                    GPIO.output(self.port, self.activate)
-                    sleep(self.duration / 1000.0)
-                    GPIO.output(self.port, self.deactivate)
-                    self.event.clear()
+        if (self.port == ConfigReader.config.conf['buzzer'] and ConfigReader.config.conf['buzzer_valid'] or
+                self.port == ConfigReader.config.conf['buzzer_next'] and
+                ConfigReader.config.conf['buzzer_next_valid']):
+            if self.nbevent > 0 or self.nbevent == -1:
+                if self.state:
+                    self.__deactivate()
+                else:
+                    self.__activate()
+                if self.nbevent > 0:
+                    self.nbevent = self.nbevent - 1
+                if self.nbevent == 0:
+                    self.stop()
 
-        # GPIO.cleanup(self.port)
-        GPIO.cleanup()
+    def __activate(self):
+        GPIO.output(self.port, self.activate)
+        self.state = True
 
-'''
+    def __deactivate(self):
+        GPIO.output(self.port, self.deactivate)
+        self.state = False
+
+
+def event_detected(port):
+    print("callback " + str(port))
+
+
 def event_detected(port):
     print("callback " + str(port))
 
 
 class rpi_gpio(QObject):
     signal_buzzer = pyqtSignal(int)
-    signal_buzzer_next = pyqtSignal()
+    signal_buzzer_next = pyqtSignal(int)
 
     def __init__(self, rpi, btn_next_action, btn_baseA, btn_baseB):
         super().__init__()
@@ -181,26 +125,15 @@ class rpi_gpio(QObject):
             if btn_baseB is not None:
                 addCallback(ConfigReader.config.conf['btn_baseB'], btn_baseB, False)
 
-    def __del__(self):
-        if self.buzzer != None:
-            self.buzzer.terminated = True
-            self.buzzer.join()
-
     def buzzer_fct(self, nb):
         print("buzzer base")
-        if self.buzzer != None:
-            self.buzzer.event.set()
-        '''
-            if nb == 1:
-                self.buzzer.event.set()
-            else:
-                self.buzzer.blink(nb)
-        '''
+        if self.buzzer is not None:
+            self.buzzer.slot_blink("blink", nb)
 
-    def buzzer_next_fct(self):
-        print("buzzer next")
-        if self.buzzer_next != None:
-            self.buzzer_next.event.set()
+    def buzzer_next_fct(self, nb):
+        print("buzzer next : ", nb)
+        if self.buzzer_next is not None:
+            self.buzzer_next.slot_blink("blink", nb)
 
 
 if __name__ == '__main__':
